Split diagnostic deduplication into its own microcrate
