+++ Cargo.toml
@@ -29,6 +29,8 @@ members = [
     "crates/perl-subprocess-runtime",
     "crates/perl-lsp-formatting-types",
     "crates/perl-lsp-formatting",
+    "crates/perl-lsp-diagnostic-types",
+    "crates/perl-lsp-diagnostic-dedup",
     "crates/perl-lsp-diagnostics",
     "crates/perl-lsp-semantic-tokens",
     "crates/perl-lsp-inlay-hints",
@@ -250,6 +252,8 @@ perl-text-line = { path = "crates/perl-text-line", version = "0.10.0" }
 perl-content-length-framing = { path = "crates/perl-content-length-framing", version = "0.10.0" }
 perl-subprocess-runtime = { path = "crates/perl-subprocess-runtime", version = "0.10.0" }
 perl-lsp-protocol = { path = "crates/perl-lsp-protocol", version = "0.10.0" }
+perl-lsp-diagnostic-types = { path = "crates/perl-lsp-diagnostic-types", version = "0.10.0" }
+perl-lsp-diagnostic-dedup = { path = "crates/perl-lsp-diagnostic-dedup", version = "0.10.0" }
 perl-uri = { path = "crates/perl-uri", version = "0.10.0" }
 perl-path-security = { path = "crates/perl-path-security", version = "0.10.0" }
 perl-workspace-discovery = { path = "crates/perl-workspace-discovery", version = "0.10.0" }
