+++ crates/perl-lsp-diagnostic-dedup/README.md
@@ -1,0 +1,3 @@
+# perl-lsp-diagnostic-dedup
+
+Shared diagnostic deduplication utilities for Perl LSP crates.
