+++ crates/perl-lsp-diagnostics/Cargo.toml
@@ -17,6 +17,8 @@ categories = ["development-tools", "text-editors"]
 doctest = false
 
 [dependencies]
+perl-lsp-diagnostic-types = { workspace = true }
+perl-lsp-diagnostic-dedup = { workspace = true }
 perl-parser-core = { workspace = true }
 perl-semantic-analyzer = { workspace = true }
 perl-workspace-index = { workspace = true }
