+++ crates/perl-lsp-diagnostics/src/diagnostics.rs
@@ -9,8 +9,7 @@ use perl_semantic_analyzer::scope_analyzer::ScopeAnalyzer;
 
 use crate::scope::scope_issues_to_diagnostics;
 
-// Re-export types from types module
-pub use crate::types::{Diagnostic, DiagnosticSeverity, DiagnosticTag, RelatedInformation};
+use perl_lsp_diagnostic_types::{Diagnostic, DiagnosticSeverity};
 
 /// Diagnostics provider
 ///
