+++ crates/perl-lsp-diagnostics/src/lib.rs
@@ -26,8 +26,6 @@
 /// Dead code detection
 #[cfg(not(target_arch = "wasm32"))]
 mod dead_code;
-/// Diagnostic deduplication utilities
-mod dedup;
 /// Core diagnostics provider
 mod diagnostics;
 /// ERROR node classification and reporting
@@ -38,13 +36,13 @@ mod lints;
 mod parse_errors;
 /// Scope analysis integration
 mod scope;
-/// Core diagnostic types
-mod types;
 /// AST walker utilities
 mod walker;
 
-pub use diagnostics::{
-    Diagnostic, DiagnosticSeverity, DiagnosticTag, DiagnosticsProvider, RelatedInformation,
+pub use diagnostics::DiagnosticsProvider;
+pub use perl_lsp_diagnostic_dedup::deduplicate_diagnostics;
+pub use perl_lsp_diagnostic_types::{
+    Diagnostic, DiagnosticSeverity, DiagnosticTag, RelatedInformation,
 };
 
 // Re-export lint checks from the lints module
