+++ crates/perl-lsp-diagnostics/src/lints/deprecated.rs
@@ -5,8 +5,10 @@
 
 use perl_parser_core::ast::{Node, NodeKind};
 
-use super::super::types::{Diagnostic, DiagnosticSeverity, DiagnosticTag, RelatedInformation};
 use super::super::walker::walk_node;
+use perl_lsp_diagnostic_types::{
+    Diagnostic, DiagnosticSeverity, DiagnosticTag, RelatedInformation,
+};
 
 /// Check for deprecated syntax
 ///
